Moved fakeFile to dedicated file
