+++ internal/filter_test.go
@@ -3,43 +3,11 @@ package internal
 import (
 	"os"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 )
 
-type fakeFile struct {
-	name    string
-	size    int64
-	modTime time.Time
-	isDir   bool
-}
-
-func (f fakeFile) Name() string {
-	return f.name
-}
-
-func (f fakeFile) Size() int64 {
-	return f.size
-}
-
-func (f fakeFile) Mode() os.FileMode {
-	panic("implement me")
-}
-
-func (f fakeFile) ModTime() time.Time {
-	return f.modTime
-}
-
-func (f fakeFile) IsDir() bool {
-	return f.isDir
-}
-
-func (f fakeFile) Sys() interface{} {
-	panic("implement me")
-}
-
-func TestApplyFilters(t *testing.T) {
+func Test_applyFilters(t *testing.T) {
 	tests := []struct {
 		name     string
 		arg      []os.FileInfo
